// src/generated/graphql.tsx
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};

export type BasicMission = {
  name?: Maybe<Scalars["String"]>;
  flight?: Maybe<Scalars["Int"]>;
};

export enum CacheControlScope {
  Public = "PUBLIC",
  Private = "PRIVATE"
}

export type Capsule = {
  capsule_serial?: Maybe<Scalars["String"]>;
  capsule_id?: Maybe<Scalars["String"]>;
  status?: Maybe<Scalars["String"]>;
  original_launch?: Maybe<Scalars["String"]>;
  original_launch_unix?: Maybe<Scalars["Float"]>;
  missions?: Maybe<Array<Maybe<BasicMission>>>;
  landings?: Maybe<Scalars["Int"]>;
  type?: Maybe<Scalars["String"]>;
  details?: Maybe<Scalars["String"]>;
  reuse_count?: Maybe<Scalars["Int"]>;
};

export enum CapsuleRange {
  Past = "past",
  Upcoming = "upcoming"
}

export type Core = {
  core_serial?: Maybe<Scalars["String"]>;
  block?: Maybe<Scalars["Int"]>;
  status?: Maybe<Scalars["String"]>;
  original_launch?: Maybe<Scalars["String"]>;
  original_launch_unix?: Maybe<Scalars["Float"]>;
  missions?: Maybe<Array<Maybe<BasicMission>>>;
  reuse_count?: Maybe<Scalars["Int"]>;
  rtls_attempts?: Maybe<Scalars["Int"]>;
  rtls_landings?: Maybe<Scalars["Int"]>;
  asds_attempts?: Maybe<Scalars["Int"]>;
  asds_landings?: Maybe<Scalars["Int"]>;
  water_landing?: Maybe<Scalars["Boolean"]>;
  details?: Maybe<Scalars["String"]>;
};

export type Dimension = {
  meters?: Maybe<Scalars["Float"]>;
  feet?: Maybe<Scalars["Float"]>;
};

export type Dragon = {
  id?: Maybe<Scalars["String"]>;
  name?: Maybe<Scalars["String"]>;
  type?: Maybe<Scalars["String"]>;
  active?: Maybe<Scalars["Boolean"]>;
  crew_capacity?: Maybe<Scalars["Int"]>;
  sidewall_angle_deg?: Maybe<Scalars["Int"]>;
  orbit_duration_yr?: Maybe<Scalars["Int"]>;
  dry_mass_kg?: Maybe<Scalars["Int"]>;
  dry_mass_lb?: Maybe<Scalars["Int"]>;
  first_flight?: Maybe<Scalars["String"]>;
  heat_shield?: Maybe<HeatShield>;
  thrusters?: Maybe<Array<Maybe<Thruster>>>;
  launch_payload_mass?: Maybe<Mass>;
  launch_payload_vol?: Maybe<PayloadVolume>;
  return_payload_mass?: Maybe<Mass>;
  return_payload_vol?: Maybe<PayloadVolume>;
  pressurized_capsule?: Maybe<PressurizedCapsule>;
  trunk?: Maybe<Trunk>;
  height_w_trunk?: Maybe<Dimension>;
  diameter?: Maybe<Dimension>;
  wikipedia?: Maybe<Scalars["String"]>;
  description?: Maybe<Scalars["String"]>;
};

export type Engines = {
  number?: Maybe<Scalars["Int"]>;
  type?: Maybe<Scalars["String"]>;
  version?: Maybe<Scalars["String"]>;
  layout?: Maybe<Scalars["String"]>;
  engine_loss_max?: Maybe<Scalars["Int"]>;
  propellant_1?: Maybe<Scalars["String"]>;
  propellant_2?: Maybe<Scalars["String"]>;
  thrust_sea_level?: Maybe<Thrust>;
  thrust_vacuum?: Maybe<Thrust>;
  thrust_to_weight?: Maybe<Scalars["Float"]>;
};
// ...
export type History = {
  id?: Maybe<Scalars["Int"]>;
  title?: Maybe<Scalars["String"]>;
  event_date_utc?: Maybe<Scalars["String"]>;
  event_date_unix?: Maybe<Scalars["Float"]>;
  flight_number?: Maybe<Scalars["Int"]>;
  details?: Maybe<Scalars["String"]>;
  links?: Maybe<HistoryLinks>;
};

export type HistoryLinks = {
  article?: Maybe<Scalars["String"]>;
  reddit?: Maybe<Scalars["String"]>;
  wikipedia?: Maybe<Scalars["String"]>;
};

export type Info = {
  name?: Maybe<Scalars["String"]>;
  founder?: Maybe<Scalars["String"]>;
  founded?: Maybe<Scalars["Int"]>;
  employees?: Maybe<Scalars["Int"]>;
  vehicles?: Maybe<Scalars["Int"]>;
  launch_sites?: Maybe<Scalars["Int"]>;
  test_sites?: Maybe<Scalars["Int"]>;
  ceo?: Maybe<Scalars["String"]>;
  cto?: Maybe<Scalars["String"]>;
  coo?: Maybe<Scalars["String"]>;
  cto_propulsion?: Maybe<Scalars["String"]>;
  valuation?: Maybe<Scalars["Float"]>;
  headquarters?: Maybe<Headquarters>;
  summary?: Maybe<Scalars["String"]>;
};

export type LandingLegs = {
  number?: Maybe<Scalars["Int"]>;
  material?: Maybe<Scalars["String"]>;
};

export type Landingpad = {
  id?: Maybe<Scalars["String"]>;
  full_name?: Maybe<Scalars["String"]>;
  status?: Maybe<Scalars["String"]>;
  location?: Maybe<Location>;
  landing_type?: Maybe<Scalars["String"]>;
  attempted_landings?: Maybe<Scalars["Int"]>;
  successful_landings?: Maybe<Scalars["Int"]>;
  wikipedia?: Maybe<Scalars["String"]>;
  details?: Maybe<Scalars["String"]>;
};

export type Launch = {
  flight_number?: Maybe<Scalars["Int"]>;
  mission_id?: Maybe<Array<Maybe<Scalars["String"]>>>;
  mission_name?: Maybe<Scalars["String"]>;
  launch_year?: Maybe<Scalars["Int"]>;
  launch_date_unix?: Maybe<Scalars["Float"]>;
  launch_date_utc?: Maybe<Scalars["String"]>;
  launch_date_local?: Maybe<Scalars["String"]>;
  is_tentative?: Maybe<Scalars["Boolean"]>;
  tentative_max_precision?: Maybe<Scalars["String"]>;
  tbd?: Maybe<Scalars["Boolean"]>;
  launch_window?: Maybe<Scalars["Int"]>;
  rocket?: Maybe<LaunchRocket>;
  ships?: Maybe<Array<Maybe<Scalars["String"]>>>;
  telemetry?: Maybe<LaunchTelemetry>;
  launch_site?: Maybe<LaunchSite>;
  launch_success?: Maybe<Scalars["Boolean"]>;
  links?: Maybe<LaunchLinks>;
  details?: Maybe<Scalars["String"]>;
  upcoming?: Maybe<Scalars["Boolean"]>;
  static_fire_date_utc?: Maybe<Scalars["String"]>;
  static_fire_date_unix?: Maybe<Scalars["Float"]>;
  timeline?: Maybe<LaunchTimeline>;
};
// ...
export type Launchpad = {
  id?: Maybe<Scalars["Int"]>;
  status?: Maybe<Scalars["String"]>;
  location?: Maybe<Location>;
  vehicles_launched?: Maybe<Array<Maybe<Scalars["String"]>>>;
  attempted_launches?: Maybe<Scalars["Int"]>;
  successful_launches?: Maybe<Scalars["Int"]>;
  wikipedia?: Maybe<Scalars["String"]>;
  details?: Maybe<Scalars["String"]>;
  site_id?: Maybe<Scalars["String"]>;
  site_name_long?: Maybe<Scalars["String"]>;
};
// ...
export type Mass = {
  kg?: Maybe<Scalars["Int"]>;
  lb?: Maybe<Scalars["Int"]>;
};

export type Mission = {
  mission_name?: Maybe<Scalars["String"]>;
  mission_id?: Maybe<Scalars["String"]>;
  manufacturers?: Maybe<Array<Maybe<Scalars["String"]>>>;
  payload_ids?: Maybe<Array<Maybe<Scalars["String"]>>>;
  wikipedia?: Maybe<Scalars["String"]>;
  website?: Maybe<Scalars["String"]>;
  twitter?: Maybe<Scalars["String"]>;
  description?: Maybe<Scalars["String"]>;
};

export enum Order {
  Asc = "asc",
  Desc = "desc"
}

export type Payload = {
  payload_id?: Maybe<Scalars["String"]>;
  norad_id?: Maybe<Array<Maybe<Scalars["String"]>>>;
  cap_serial?: Maybe<Scalars["String"]>;
  reused?: Maybe<Scalars["Boolean"]>;
  customers?: Maybe<Array<Maybe<Scalars["String"]>>>;
  nationality?: Maybe<Scalars["String"]>;
  manufacturer?: Maybe<Scalars["String"]>;
  payload_type?: Maybe<Scalars["String"]>;
  payload_mass_kg?: Maybe<Scalars["Float"]>;
  payload_mass_lbs?: Maybe<Scalars["Float"]>;
  orbit?: Maybe<Scalars["String"]>;
  orbit_params?: Maybe<PayloadOrbitParams>;
  mass_returned_kg?: Maybe<Scalars["Float"]>;
  mass_returned_lbs?: Maybe<Scalars["Float"]>;
  flight_time_sec?: Maybe<Scalars["Int"]>;
  cargo_manifest?: Maybe<Scalars["String"]>;
};
// ...
export type Query = {
  capsule?: Maybe<Capsule>;
  capsules?: Maybe<Array<Maybe<Capsule>>>;
  core?: Maybe<Core>;
  cores?: Maybe<Array<Maybe<Core>>>;
  dragon?: Maybe<Dragon>;
  dragons?: Maybe<Array<Maybe<Dragon>>>;
  history?: Maybe<Array<Maybe<History>>>;
  info?: Maybe<Info>;
  landingpad?: Maybe<Landingpad>;
  landingpads?: Maybe<Array<Maybe<Landingpad>>>;
  launch?: Maybe<Launch>;
  launches?: Maybe<Array<Maybe<Launch>>>;
  launchpad?: Maybe<Launchpad>;
  launchpads?: Maybe<Array<Maybe<Launchpad>>>;
  mission?: Maybe<Mission>;
  missions?: Maybe<Array<Maybe<Mission>>>;
  payload?: Maybe<Payload>;
  payloads?: Maybe<Array<Maybe<Payload>>>;
  roadster?: Maybe<Roadster>;
  rocket?: Maybe<Rocket>;
  rockets?: Maybe<Array<Maybe<Rocket>>>;
  ship?: Maybe<Ship>;
  ships?: Maybe<Array<Maybe<Ship>>>;
};
// ...
export type Roadster = {
  name: Scalars["String"];
  launch_date_utc?: Maybe<Scalars["String"]>;
  launch_date_unix?: Maybe<Scalars["Float"]>;
  launch_mass_kg?: Maybe<Scalars["Int"]>;
  launch_mass_lbs?: Maybe<Scalars["Int"]>;
  norad_id?: Maybe<Scalars["Int"]>;
  epoch_jd?: Maybe<Scalars["Float"]>;
  orbit_type?: Maybe<Scalars["String"]>;
  apoapsis_au?: Maybe<Scalars["Float"]>;
  semi_major_axis_au?: Maybe<Scalars["Float"]>;
  eccentricity?: Maybe<Scalars["Float"]>;
  inclination?: Maybe<Scalars["Float"]>;
  longitude?: Maybe<Scalars["Float"]>;
  periapsis_arg?: Maybe<Scalars["Float"]>;
  period_days?: Maybe<Scalars["Float"]>;
  speed_kph?: Maybe<Scalars["Float"]>;
  speed_mph?: Maybe<Scalars["Float"]>;
  earth_distance_km?: Maybe<Scalars["Float"]>;
  earth_distance_mi?: Maybe<Scalars["Float"]>;
  mars_distance_km?: Maybe<Scalars["Float"]>;
  mars_distance_mi?: Maybe<Scalars["Float"]>;
  wikipedia?: Maybe<Scalars["String"]>;
  details?: Maybe<Scalars["String"]>;
};

export type Rocket = {
  id?: Maybe<Scalars["Int"]>;
  active?: Maybe<Scalars["Boolean"]>;
  stages?: Maybe<Scalars["Int"]>;
  boosters?: Maybe<Scalars["Int"]>;
  cost_per_launch?: Maybe<Scalars["Int"]>;
  success_rate_pct?: Maybe<Scalars["Float"]>;
  first_flight?: Maybe<Scalars["String"]>;
  country?: Maybe<Scalars["String"]>;
  company?: Maybe<Scalars["String"]>;
  height?: Maybe<Dimension>;
  diameter?: Maybe<Dimension>;
  mass?: Maybe<Mass>;
  payload_weights?: Maybe<Array<Maybe<RocketPayloadWeight>>>;
  first_stage?: Maybe<RocketFirstStage>;
  second_stage?: Maybe<RocketSecondStage>;
  engines?: Maybe<Engines>;
  landing_legs?: Maybe<LandingLegs>;
  flickr_images?: Maybe<Array<Maybe<Scalars["String"]>>>;
  wikipedia?: Maybe<Scalars["String"]>;
  description?: Maybe<Scalars["String"]>;
  rocket_id?: Maybe<Scalars["String"]>;
  rocket_name?: Maybe<Scalars["String"]>;
  rocket_type?: Maybe<Scalars["String"]>;
};

export type RocketFirstStage = {
  reusable?: Maybe<Scalars["Boolean"]>;
  engines?: Maybe<Scalars["Int"]>;
  fuel_amount_tons?: Maybe<Scalars["Float"]>;
  burn_time_sec?: Maybe<Scalars["Int"]>;
  thrust_sea_level?: Maybe<Thrust>;
  thrust_vacuum?: Maybe<Thrust>;
};

export type RocketPayload = {
  option_1?: Maybe<Scalars["String"]>;
  option_2?: Maybe<Scalars["String"]>;
  composite_fairing?: Maybe<Fairing>;
};

export type RocketPayloadWeight = {
  id?: Maybe<Scalars["String"]>;
  name?: Maybe<Scalars["String"]>;
  kg?: Maybe<Scalars["Float"]>;
  lb?: Maybe<Scalars["Float"]>;
};

export type RocketSecondStage = {
  engines?: Maybe<Scalars["Int"]>;
  fuel_amount_tons?: Maybe<Scalars["Float"]>;
  burn_time_sec?: Maybe<Scalars["Int"]>;
  thrust?: Maybe<Thrust>;
  payloads?: Maybe<RocketPayload>;
};

export type Ship = {
  ship_id?: Maybe<Scalars["String"]>;
  ship_name?: Maybe<Scalars["String"]>;
  ship_model?: Maybe<Scalars["String"]>;
  ship_type?: Maybe<Scalars["String"]>;
  roles?: Maybe<Array<Maybe<Scalars["String"]>>>;
  active?: Maybe<Scalars["Boolean"]>;
  imo?: Maybe<Scalars["Int"]>;
  mmsi?: Maybe<Scalars["Int"]>;
  abs?: Maybe<Scalars["Int"]>;
  class?: Maybe<Scalars["Int"]>;
  weight_lbs?: Maybe<Scalars["Float"]>;
  weight_kg?: Maybe<Scalars["Float"]>;
  year_built?: Maybe<Scalars["Int"]>;
  home_port?: Maybe<Scalars["String"]>;
  status?: Maybe<Scalars["String"]>;
  speed_kn?: Maybe<Scalars["Float"]>;
  course_deg?: Maybe<Scalars["Float"]>;
  position?: Maybe<Position>;
  successful_landings?: Maybe<Scalars["Int"]>;
  attempted_landings?: Maybe<Scalars["Int"]>;
  missions?: Maybe<Array<Maybe<BasicMission>>>;
  url?: Maybe<Scalars["String"]>;
  image?: Maybe<Scalars["String"]>;
};

export type Thrust = {
  kN?: Maybe<Scalars["Float"]>;
  lbf?: Maybe<Scalars["Float"]>;
};
// ...
export type RocketQuery = { __typename?: "Query" } & {
  rocket: Maybe<
    { __typename?: "Rocket" } & Pick<
      Rocket,
      | "description"
      | "rocket_name"
      | "rocket_type"
      | "first_flight"
      | "flickr_images"
      | "country"
      | "company"
    > & {
        first_stage: Maybe<
          { __typename?: "RocketFirstStage" } & Pick<
            RocketFirstStage,
            "fuel_amount_tons"
          >
        >;
        second_stage: Maybe<
          { __typename?: "RocketSecondStage" } & Pick<
            RocketSecondStage,
            "fuel_amount_tons"
          >
        >;
        height: Maybe<{ __typename?: "Dimension" } & Pick<Dimension, "meters">>;
        diameter: Maybe<
          { __typename?: "Dimension" } & Pick<Dimension, "meters">
        >;
        mass: Maybe<{ __typename?: "Mass" } & Pick<Mass, "kg">>;
      }
  >;
};